Update Index.rst
I have proposed some major changes on this page. This tutorial is an excellent starting point for those familiar enough with HTML specifically and the whole concept of markup language in general. I was able to have a better breakthrough with this tutorial than any other before.
I also see the possibility for more improvement in the flow of the teaching. There are parts that I could understand easily, but there are parts where there needs to be some more clarification in steps. 
In this page I took the liberty to seek a more uniform use of verbs and phrase constructions so that the reader doesn´t have to readjust his/her mind for the changes in expressions and construction. An example of this is the transformation I applied to the note. I turned the various phrases into subitems of a list. I also unified the first gerunds of the last list on the different possibilities by applying the verb "using". I believe that the readers will filter out the repetition and focus on the main purpose of the phrases.
I am available for any doubt clarification,