Update Index.rst
I am not understanding the meaning of this phrase, "However, we will also repeat concepts in the Templating Tutorial where they are needed." Are we referring to the same tutorial or another one?
I would like to suggest inserting github as a means to help with the "issues" in the documentation. I only see email and forge as the two means for suggestions. I am enjoying working with github as I can review my corrections before I send the pull request.