Update Index.rst
OK,
Oh, wow! I am really enjoying this because I am learning very much.
I did a lot of editing in this particular page.  I have tried to fill in details where I think a newcomer would get lost. I have also tried adding in a link to the Introduction Package so that newcomers will, at least, be able to go there and see what it is. This idea of linking, I think, will really help in the learning process as each one can access the info if they are not clear on what is being said. Could you point me to some kind of link that will teach me how to create links in GigHub? 
I have quite a few comments on some of the steps I reviewed. 

I have chosen pop-up (not popup as it would sound confusing) and make a distinction between our TYPO3 context menu and the browser pop-up. There may need to be done some more reviewing so that there can be better clarity on the two.
An issue that may confuse the "uninitiated" (like me :) ) is the "name" Root and the concept "root page".  It might not be a problem in this tutorial, but thinking of future tutorials, this could be a source of confusion. I would like to suggest a different name for the root page (i.e. Main, LandingPage...., etc.).
Step 1. I would like to suggest more indicators (blue squares or rectangles, arrows in red, numbered sequences, etc.) so that the readers can get the flow of clicks and steps in the instructions. In this particular step, I don´t see a major problem, but it prepares the learner for steps up ahead, such as in step 4, where they could get lost.
Step 2 ibid.
Step 3 (see comment at beginning on name the root page, "Root").
Step 4 (see comment in Step 1)
Step 5 could use an indicator in the image (a blue square as used in the "save and close" icon image for step 15) for the newbie to understand what the "Save and close document" is and means. Is there any way I could either edit the images or create new ones as a suggestion? I think editing the originals would be better as there could rise differences from different installations. How do I do this in GitHub?
Step 6. (see comment in Step 1).
Step 7. (sequence of clicks as in Step 1).
Step 8. There is missing the illustration as to how to create the name "Home" for the first page created. I know the idea was convered in creatign the "Root" root page, but it still might not have been internalized by the learner. I added extra instructions in case the new user will get lost. The term "palette" came up once in Step 4. I think it would be nice to use it more often so that the term can be uderstood better in the tutorials. I, for one, only understood what a palette is in this tutorial. 
Step 9. no comment.
Step10. no comment.
Step 11 could confuse a newbie as they will see the "Imprint" page selected under the context menu. So, I chose to add more detail just in case. Once again, more details in the image could help (comment in Step 1).
Step 12. Added more comments so the learner can "see" what is happening.
Step 13. Added a little more detail. I chose, again, to clearly state that the browser pop-up menu is going to appear.
Step 14. more detail binding previous steps (...we named "Home"...) to what is going to happen after (i.e., the disappearance of the pop-up menu).
Step 15. (see comment in Step 5 on the "save and close document" icon)
I guess that is it for now. 
wish you the best,
Tim