Update Index.rst
I have tried to make the language smoother for readers and also added in some more ideas and suggestions (such as repeating the terms "markers" and "subparts"). 
I see the need to define better the term "content" better, but I am not sure if I should add it here in this tutorial, or if I should consider creating a kind of glossary tutorial. Another idea I have for this is to create links to any kind of definition or explanation to these terms.
Tim