Update Index.rst
I have tried to add some more clarity, especially with the differences between markers and subparts. I know there is more to do in this chapter, especially in the last part with the extra add-ins of the markers and subparts. It gets a little confusing when some of the markers and subparts are added into the HTML template (i.e., where exactly to insert <div id="headertitle">###HEADERTITLE###</div>, and <div id="breadcrumb">###DATE###&nbsp;:::&nbsp;###BREADCRUMB###</div> is also confusing at first).
But I believe, that the greatest source of confusion has been the lack of comprehension between what a marker and a subpart are.
I hope I am not overdoing it.
Tim