[TASK] Clear up usage of marks and subparts
The tutorial does not define clearly enough what is a mark,
a subpart and a marker, causing confusion for new users.
The main chapter about marks and subparts is reworked to improve
the learning experience.

Furthermore the reST markup is also improved in a number
of places. "TYPO3" is also replaced by "TYPO3 CMS" in many
places.